# Faster and simpler filter implementation