… bench/785.vector.pbf
… bench/filters.json