Closes #10, ref #8. Alternative approach, still using eval.

Filters have the same performance, but filter creation is 40% faster, and there is 30% less code. 

It also makes `in` with a lot of items faster due to using `indexOf`, and it'll be easier to significantly improve further on this basis.

cc @jfirebaugh @dcervelli