@mourner 2016-01-29
Benchmarks indicate that it works much faster for simpler filters and on par with #9 on huge filters: 

```
# master
create 5-item filter: 1.174ms
filter 1 million times: 89.382ms
create 64000-item filter: 345.341ms
filter 1 million times: 726.251ms

# pr 9
create 5-item filter: 1.081ms
filter 1 million times: 89.812ms
create 64000-item filter: 290.508ms
gfilter 1 million times: 132.900ms

# simple 
create 5-item filter: 0.796ms
filter 1 million times: 30.522ms
create 64000-item filter: 28.999ms
filter 1 million times: 164.089ms
```