@mourner 2016-01-29
Disregard the 64k results, had a faulty benchmark. Here is a correct benchmark for 2k:

```
# master 
create 2000-item filter: 10.961ms
filter 1 million times: 20915.215ms

# pr 9
create 2000-item filter: 6.585ms
filter 1 million times: 215.851ms

# simple
create 2000-item filter: 0.747ms
filter 1 million times: 1414.595ms
```

So for filtering against a huge number of items, this branch is much faster than master but much slower than pr/9. I'll now try to implement binary search and this should change things dramatically though.