@mourner 2016-01-29
Added binary search! It turns on for more than 200 items (picked by testing running benchmarks). Results are great:

```
# pr 9
create 100000-item filter: 428.668ms
filter 1 million times: 499.984ms

# simple
create 100000-item filter: 71.494ms
filter 1 million times: 383.677ms
```