@dcervelli 2016-01-29
For random string data (which is closer to our use case) I get these results:

```
# pr 9
create 64000-item filter: 240.739ms
filter 1 million times: 757.678ms

# simple
create 64000-item filter: 161.316ms
filter 1 million times: 1879.914ms
```

Given the simpler code, I'm totally OK with this until it proves to be a bottleneck (which I doubt it will).

And for calibration: on my box for bench/big.js I get:

```
# simple
create 64000-item filter: 38.474ms
filter 1 million times: 324.140ms
```

Here's my benchmark:

``` javascript
'use strict';

var filter = require('../');

var N = 64000;

var arr = ['in', 'foo'];
for (var i = 0; i < N; i++) arr.push(i + " " + Math.random());

console.time('create ' + N + '-item filter');
var f = filter(arr);
console.timeEnd('create ' + N + '-item filter');

var feature = {properties: {foo: 0}};

console.time('filter 1 million times');
for (var i = 0; i < 1000000; i++) {
    feature.properties.foo = arr[Math.floor(Math.random() * N) + 2];
    f(feature);
}
console.timeEnd('filter 1 million times');

```

Thanks!